pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "faiss-cpu>=1.10.0",
    "langchain>=0.3.23",
    "numpy>=2.2.4",
    "openai>=1.70.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "spacy>=3.8.5",
    "streamlit>=1.44.1",
]
